requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
tqdm
google-generativeai